+++ packages/client/hmi-client/src/components/widgets/tera-chart-settings.vue
@@ -30,7 +30,6 @@
 		<tera-chart-control
 			v-if="type !== ChartSettingType.VARIABLE_COMPARISON && type !== ChartSettingType.SENSITIVITY"
 			:chart-config="{
-				selectedRun: 'fixme',
 				selectedVariable: selectedOptions ?? []
 			}"
 			:multi-select="true"
@@ -42,66 +41,15 @@
 		<slot></slot>
 
 		<!-- Sensitivity analysis settings -->
-		<template v-if="type === ChartSettingType.SENSITIVITY && sensitivityOptions">
-			<!--FIXME: It might be better to move these inside the panel so that they can be controlled at an individual chart settings level -->
-
-			<div class="mad-libs-container">
-				Consider how sensitive the
-				<Dropdown
-					:disabled="_.isEmpty(selectedOptions)"
-					placeholder="Select method"
-					:model-value="sensitivityOptions.method"
-					:options="Object.values(SensitivityMethod)"
-					@change="onUpdateSensitivitySettings('method', $event.value, sensitivityOptions)"
-				/>
-				<tera-input-number
-					v-if="sensitivityOptions.method === SensitivityMethod.TIMEPOINT"
-					:disabled="_.isEmpty(selectedOptions)"
-					:model-value="sensitivityOptions.timepoint"
-					auto-width
-					@update:model-value="onUpdateSensitivitySettings('timepoint', $event, sensitivityOptions)"
-				/>
-				of the outcome(s) of interest
-				<tera-chart-control
-					class="overflow-hidden"
-					:chart-config="{
-						selectedRun: 'fixme',
-						selectedVariable: selectedOptions ?? []
-					}"
-					:multi-select="true"
-					:show-remove-button="false"
-					:variables="selectOptions"
-					@configuration-change="$emit('selection-change', $event.selectedVariable, type)"
-				/>
-				is to the model parameter(s)
-				<MultiSelect
-					class="overflow-hidden"
-					:disabled="_.isEmpty(selectedOptions)"
-					placeholder="Select parameters"
-					:model-value="sensitivityOptions.selectedInputVariables"
-					:options="sensitivityOptions.inputOptions"
-					@change="onUpdateSensitivitySettings('selectedInputVariables', $event.value, sensitivityOptions)"
-					filter
-				>
-					<template v-slot:value>
-						<template v-for="(variable, index) in sensitivityOptions.selectedInputVariables" :key="index">
-							<template v-if="index > 0">,&nbsp;</template>
-							<span> {{ variable }} </span>
-						</template>
-					</template>
-				</MultiSelect>
-			</div>
-			<div v-for="option in sensitivityChartOptions" class="flex align-items-center gap-2 mt-1" :key="option.value">
-				<RadioButton
-					:disabled="_.isEmpty(selectedOptions)"
-					:model-value="sensitivityOptions.chartType"
-					:value="option.value"
-					name="sensitivityChartTypes"
-					@change="onUpdateSensitivitySettings('chartType', option.value, sensitivityOptions)"
-				/>
-				<label :class="{ disabled: isEmpty(selectedOptions) }" :for="option.value">{{ option.label }}</label>
-			</div>
-		</template>
+		<tera-sensitivity-chart-settings
+			v-if="type === ChartSettingType.SENSITIVITY && sensitivityOptions"
+			:settings="settings"
+			:selectOptions="selectOptions"
+			:selectedOptions="selectedOptions"
+			:sensitivityOptions="sensitivityOptions"
+			@selection-change="$emit('selection-change', $event, type)"
+			@sensitivity-selection-change="$emit('sensitivity-selection-change', $event)"
+		/>
 		<template v-if="type === ChartSettingType.VARIABLE_COMPARISON">
 			<section>
 				<tera-chart-settings-item-comparison
@@ -133,14 +81,11 @@ import TeraCheckbox from '@/components/widgets/tera-checkbox.vue';
 import TeraChartSettingsItem from '@/components/widgets/tera-chart-settings-item.vue';
 import TeraChartSettingsItemComparison from '@/components/widgets/tera-chart-settings-item-comparison.vue';
 import TeraChartControl from '@/components/workflow/tera-chart-control.vue';
-import MultiSelect from 'primevue/multiselect';
 import { ChartSetting, ChartSettingType, SensitivityChartType, SensitivityMethod } from '@/types/common';
 import { computed } from 'vue';
 import { EnsembleVariableChartSettingOption, getEnsembleChartSettingOptions } from '@/services/chart-settings';
-import _, { isEmpty } from 'lodash';
-import RadioButton from 'primevue/radiobutton';
-import Dropdown from 'primevue/dropdown';
-import TeraInputNumber from './tera-input-number.vue';
+import _ from 'lodash';
+import teraSensitivityChartSettings from './tera-sensitivity-chart-settings.vue';
 
 const props = defineProps<{
 	title?: string; // Optional title for the settings panel
@@ -176,11 +121,6 @@ const emits = defineEmits([
 	'comparison-selection-change'
 ]);
 
-const sensitivityChartOptions = [
-	{ label: 'Scatter', value: SensitivityChartType.SCATTER },
-	{ label: 'Heatmap', value: SensitivityChartType.HEATMAP }
-];
-
 // Settings of the same type that we want to interact with.
 const targetSettings = computed(() => props.settings.filter((s) => s.type === props.type));
 
@@ -188,13 +128,6 @@ const targetSettings = computed(() => props.settings.filter((s) => s.type === pr
 const ensembleChartOptions = computed(() => getEnsembleChartSettingOptions(targetSettings.value));
 const toggleEnsembleChartOption = (option: EnsembleVariableChartSettingOption, value: boolean) => {
 	emits('toggle-ensemble-variable-setting-option', option, value);
-};
-
-const onUpdateSensitivitySettings = (key: string, value: any, sensitivityOptions: typeof props.sensitivityOptions) => {
-	emits('sensitivity-selection-change', {
-		...sensitivityOptions,
-		[key]: value
-	});
 };
 </script>
 <style scoped>
@@ -227,30 +160,5 @@ const onUpdateSensitivitySettings = (key: string, value: any, sensitivityOptions
 .actions {
 	display: flex;
 	gap: var(--gap-2);
-}
-
-.mad-libs-container {
-	display: flex;
-	flex-wrap: wrap;
-	align-items: center;
-	gap: var(--gap-2);
-	border: 1px solid var(--gray-300);
-	border-radius: var(--border-radius);
-	padding: var(--gap-2);
-	border-left: 4px solid var(--primary-color);
-	background: var(--surface-0);
-	box-shadow: 0 2px 4px -1px rgba(0, 0, 0, 0.08);
-	overflow: hidden;
-}
-
-:deep(.p-multiselect),
-:deep(.p-dropdown) {
-	.p-multiselect-label,
-	.p-dropdown-label {
-		padding: var(--gap-1);
-		white-space: nowrap;
-		overflow: hidden;
-		text-overflow: ellipsis;
-	}
 }
 </style>
